pyproject.toml
[project]
name = "slidescan"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "opencv-python>=4.12.0.88",
    "pillow>=11.3.0",
    "pyserial>=3.5",
    "tk>=0.1.0",
]
